fix overflows

+++ src/api.rs
@@ -9,6 +9,7 @@ use eyre::{eyre, OptionExt, Result};
 use redis::{Client, Commands};
 use serde::{Deserialize, Serialize};
 use shared::coin::Pair;
+use tracing::warn;
 
 use crate::{
 	datapoint::{Datapoint, TimeType},
@@ -51,7 +52,7 @@ fn get_prices(
 
 	// apply corrections to amount and offset to prepare for querrying with redis
 	let offset = (offset * -1) as isize;
-	let amount = (amount * (interval / collection_interval_minutes)) as isize * -1;
+	let amount = (amount as isize * (interval / collection_interval_minutes) as isize) * -1;
 
 	let current_timestamp: i64 = connection.lindex(format!("{}:timestamps", pair.to_string()), -1)?;
